pipe_damage_detection/requirements.txt
torch
torchvision
flask
scikit-learn
pillow
matplotlib